Create observation_data.py
wergeld was interested in this code for consideration with his work. This is code snippet from a pipeline process I use regularly in my supernova searches. intent was to just upload to scripts for anyone to use for what ever purpose. project or otherwise.  Please let me know if this should not have been done. New to github usage./ Ken